[Feat]공지사항 팝업뷰 구현
